refactor(frontend): Move the start components into their own start directory
