+++ frontend/src/lib/components/start/Year.svelte
